requirements.txt
# HC-MARL Requirements
# Core dependencies for Phase 1 (core modules)

# Numerical computing
numpy>=1.24.0
scipy>=1.10.0

# QP solver for ECBF safety filter
cvxpy>=1.4.0
osqp>=0.6.3

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging and visualization (used from Phase 2 onward)
matplotlib>=3.7.0
wandb>=0.16.0

# RL framework (required for training, tests, and all agent code)
# NOTE: On GPU hosts (CUDA), install torch from the CUDA wheel index BEFORE
# running `pip install -r requirements.txt` so the line below is satisfied
# by the CUDA build (CPU wheel would otherwise be fetched from PyPI):
#   pip install torch --index-url https://download.pytorch.org/whl/cu124
# cu124 wheels are forward-compatible with CUDA 12.5/12.6/12.7/12.8 drivers.
# Minimum 2.6.0 is required for `weights_only=False` resume behaviour used
# by every agent loader (see 2026-04-17 execution hardening).
torch>=2.6.0
gymnasium>=0.29.0
# NOTE: pettingzoo is intentionally NOT a runtime dep. hcmarl.envs.pettingzoo_wrapper
# implements the PettingZoo parallel-env API surface without importing the library.

# Real benchmark environments (Phase 4B: legitimacy upgrade)
# rware>=1.0.0                    # RWARE multi-agent warehouse benchmark
# safety-gymnasium>=1.0.0         # Real Safety-Gymnasium (PKU-Alignment)
# omnisafe>=0.6.0                 # OmniSafe safe RL algorithms library

# Real dataset loaders (Phase 4B: legitimacy upgrade)
# h5py>=3.8.0                     # For RoboMimic HDF5 files
# d4rl>=1.1                       # For D4RL Adroit human demos
# minari>=0.4.0                   # Modern D4RL replacement (Farama)
